requirements.txt
requests~=2.25.0
boto3~=1.16.43
gpval
wheel
setuptools~=50.3.2
botocore~=1.19.43
PyJWT
paramiko
shapely
pandas
psycopg2-binary=2.8.6